+++ tests/test_strategy.py
